+++ src/CMakeLists.txt
@@ -67,7 +67,7 @@ set(SOURCES
     cpprealm/internal/bridge/timestamp.cpp
     cpprealm/internal/bridge/uuid.cpp
     cpprealm/internal/scheduler/realm_core_scheduler.cpp
-    cpprealm/schedulers/default_schedulers.cpp
+    cpprealm/schedulers/default_scheduler.cpp
     cpprealm/logger.cpp
     cpprealm/sdk.cpp) # REALM_SOURCES
 
@@ -127,7 +127,7 @@ set(HEADERS
     ../include/cpprealm/internal/generic_network_transport.hpp
     ../include/cpprealm/internal/type_info.hpp
     ../include/cpprealm/internal/scheduler/realm_core_scheduler.hpp
-    ../include/cpprealm/schedulers/default_schedulers.hpp
+    ../include/cpprealm/schedulers/default_scheduler.hpp
     ../include/cpprealm/logger.hpp
     ../include/cpprealm/notifications.hpp
     ../include/cpprealm/rbool.hpp
